refactor(yt/regist): moved most Dto oriented class to Dto folder
